fix(#712): move heartbeat_buffer.py from core/ to services/agents/ — not a kernel primitive
HeartbeatBuffer is agent service infrastructure (extracted from AgentRegistry per SRP),
not a kernel primitive. Moved to nexus.services.agents.heartbeat_buffer and updated
all imports across source and test files.
