+++ tests/e2e/server/test_agent_registry_e2e.py
@@ -346,7 +346,7 @@ class TestBridgeReliabilityE2E:
         for i in range(10):
             small_registry.register(f"hb-warn-{i}", "alice")
 
-        with caplog.at_level(logging.WARNING, logger="nexus.core.heartbeat_buffer"):
+        with caplog.at_level(logging.WARNING, logger="nexus.services.agents.heartbeat_buffer"):
             for i in range(10):
                 small_registry.heartbeat(f"hb-warn-{i}")
