+++ tests/unit/core/test_agent_registry.py
@@ -606,14 +606,14 @@ class TestHeartbeatCapacityWarning:
             reg.transition(f"agent-{i}", AgentState.CONNECTED, expected_generation=0)
 
         # Heartbeat 7 agents (below threshold)
-        with caplog.at_level(logging.WARNING, logger="nexus.core.heartbeat_buffer"):
+        with caplog.at_level(logging.WARNING, logger="nexus.services.agents.heartbeat_buffer"):
             caplog.clear()
             for i in range(7):
                 reg.heartbeat(f"agent-{i}")
             assert "capacity" not in caplog.text
 
         # Heartbeat the 8th agent (hits 80%)
-        with caplog.at_level(logging.WARNING, logger="nexus.core.heartbeat_buffer"):
+        with caplog.at_level(logging.WARNING, logger="nexus.services.agents.heartbeat_buffer"):
             caplog.clear()
             reg.heartbeat("agent-7")
             assert "capacity" in caplog.text
